+++ src/runners/vanilla-node/utils.ts
@@ -110,29 +110,3 @@ export const getTemplate = async (name?: string): Promise<Template> => {
   console.log(`Creating project with template: ${yellow(template)} \n`);
   return template;
 };
-
-/*
- *
- */
-export const getSuccessString = (projectName: string, template: string) => {
-  const emoji = '🦉';
-  const chars = 53;
-  const extraChars = projectName.length + template.length;
-  const hashString = Array.from({ length: extraChars + chars })
-    .map(el => '#')
-    .join('');
-
-  const successString = `
-${cyan(hashString)}
-${emoji} ${green('Success!')} Created new project ${yellow.bold(
-    projectName
-  )} using template: ${yellow.italic(template)} ${emoji}
-${cyan(hashString)}
-`;
-  return successString;
-};
-
-/* Test */
-// const projectName = 'my-new-project';
-// const template = 'node-basic';
-// console.log(getSuccessString(projectName, template));
